splitted_colored
