+++ Reader.h
@@ -1,0 +1,24 @@
+#ifndef READER_H
+#define READER_H
+
+#include <cctype>
+#include <sstream>
+
+#include "objects.h"
+
+
+class Reader{
+public:
+    Reader(); 
+    void ReadRectangle(Shapes::Rectangle* rect, XMLElement* E);
+    void ReadLine(Shapes::Line* line, XMLElement* E);
+    void ReadCircle(Shapes::Circle* circle, XMLElement* E);
+    void ReadEllipse(Shapes::Ellipse* ellipse, XMLElement* E);
+    void ReadPolygon(Shapes::Polygon* polygon, XMLElement* E);
+    void ReadPolyline(Shapes::Polyline* polyline, XMLElement* E);
+    void ReadText(Shapes::Text* text, XMLElement* E);
+    void ReadPath(Shapes::Path* path, XMLElement *E);
+};
+
+
+#endif
